+++ brdr/__init__.py
@@ -6,4 +6,4 @@ logging.basicConfig(
     datefmt="%d-%b-%y %H:%M:%S",
 )
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"