+++ brdr/aligner.py
@@ -46,8 +46,6 @@ from brdr.utils import diffs_from_dict_series, multipolygons_to_singles
 from brdr.utils import geojson_from_dict
 from brdr.utils import get_breakpoints_zerostreak
 from brdr.utils import get_series_geojson_dict
-from brdr.utils import merge_dict
-from brdr.utils import merge_dict_series
 from brdr.utils import merge_process_results
 from brdr.utils import write_geojson
 
@@ -298,7 +296,7 @@ class Aligner:
         if self.multi_as_single_modus:
             dict_thematic = multipolygons_to_singles(dict_thematic)
         for key in dict_thematic:
-            self.logger.feedback_info("thematic id to process: " + str(key))
+            self.logger.feedback_debug("thematic id to process: " + str(key))
             try:
                 dict_result[key] = self.process_geometry(
                     dict_thematic[key],
@@ -389,9 +387,9 @@ class Aligner:
             threshold_overlap_percentage=threshold_overlap_percentage,
         )
         dict_thematic = self.dict_thematic
-        if self.multi_as_single_modus:
-            dict_series = merge_dict_series(dict_series)
-            dict_thematic = merge_dict(self.dict_thematic)
+        # if self.multi_as_single_modus:
+        #     dict_series = merge_dict_series(dict_series)
+        #     dict_thematic = merge_dict(self.dict_thematic)
 
         diffs_dict = diffs_from_dict_series(dict_series, dict_thematic)
 
@@ -526,7 +524,12 @@ class Aligner:
                     last_version_date = version_date
                 if version_date is not None and version_date > last_version_date:
                     last_version_date = version_date
-            if equals(geom_intersection, geom_reference):
+
+            try:
+                equal = equals(geom_intersection, geom_reference)
+            except:
+                equal = False
+            if equal:
                 full = True
                 area = round(geom_reference.area, 2)
                 perc = 100
@@ -557,8 +560,14 @@ class Aligner:
             }
         dict_formula["full"] = full_total
         if last_version_date is not None:
-            dict_formula["versiondate"] = last_version_date.strftime(date_format)
-        geom_od = safe_difference(geometry, make_valid(unary_union(intersected)))
+            dict_formula["last_version_date"] = last_version_date.strftime(date_format)
+        geom_od = buffer_pos(
+            buffer_neg(
+                safe_difference(geometry, make_valid(unary_union(intersected))),
+                CORR_DISTANCE,
+            ),
+            CORR_DISTANCE,
+        )
         if geom_od is not None:
             area_od = round(geom_od.area, 2)
             if area_od > 0:
@@ -573,8 +582,8 @@ class Aligner:
         """
         get a dict of the results
         """
-        if self.multi_as_single_modus:
-            return merge_process_results(self.dict_result)
+        # if self.multi_as_single_modus:
+        #     return merge_process_results(self.dict_result)
         return self.dict_result
 
     def get_results_as_geojson(self, formula=False):
@@ -586,8 +595,8 @@ class Aligner:
                 in the output. Defaults to False.
         """
         results_dict = self.dict_result
-        if self.multi_as_single_modus:
-            results_dict = merge_process_results(results_dict)
+        # if self.multi_as_single_modus:
+        #     results_dict = merge_process_results(results_dict)
 
         return self.get_predictions_as_geojson(
             formula,