+++ brdr/grb.py
@@ -27,7 +27,6 @@ from brdr.utils import dict_series_by_keys
 from brdr.utils import geojson_to_dicts
 from brdr.utils import get_collection
 from brdr.utils import get_collection_by_partition
-from brdr.utils import merge_dict
 
 log = logging.getLogger(__name__)
 date_format = "%Y-%m-%d"
@@ -98,8 +97,8 @@ def get_geoms_affected_by_grb_change(
 
     """
     dict_thematic = aligner.dict_thematic
-    if aligner.multi_as_single_modus:
-        dict_thematic = merge_dict(dict_thematic)
+    # if aligner.multi_as_single_modus:
+    #     dict_thematic = merge_dict(dict_thematic)
     crs = aligner.CRS
     affected_dict: dict[str, BaseGeometry] = {}
     unchanged_dict: dict[str, BaseGeometry] = {}
@@ -294,18 +293,20 @@ def evaluate(
         for theme_id in dict_unchanged.keys():
             prop_dictionary[dist][theme_id] = {}
 
-    equality = False
-    for dist, dict_predicted_for_dist in dict_predicted.items():
-        if equality:
-            break
-        for theme_id in dict_predicted_for_dist.keys():
-            results = dict_series[dist][theme_id]
-            actual_formula = actual_aligner.get_formula(results["result"])
+    dict_predicted_keys = dict_series_by_keys(dict_predicted)
+
+    for theme_id, dist_dict in dict_predicted_keys.items():
+        equality = False
+        for dist in sorted(dist_dict.keys()):
+            if equality:
+                break
+            geomresult = dist_dict[dist][theme_id]["result"]
+            actual_formula = actual_aligner.get_formula(geomresult)
             prop_dictionary[dist][theme_id]["formula"] = json.dumps(actual_formula)
             base_formula = None
             if theme_id in thematic_dict_formula:
                 base_formula = thematic_dict_formula[theme_id]
-            equality, _property = check_equality(
+            equality, prop = check_equality(
                 base_formula,
                 actual_formula,
                 threshold_area,
@@ -313,10 +314,9 @@ def evaluate(
             )
             if equality:
                 dict_evaluated_result[dist][theme_id] = dict_predicted[dist][theme_id]
-                prop_dictionary[dist][theme_id]["evaluation"] = _property
+                prop_dictionary[dist][theme_id]["evaluation"] = prop
                 break
 
-    dict_predicted_keys = dict_series_by_keys(dict_predicted)
     evaluated_theme_ids = list(dict_series_by_keys(dict_evaluated_result).keys())
     # fill where no equality is found/ The biggest predicted distance is returned as
     # proposal
@@ -330,21 +330,14 @@ def evaluate(
                 )
                 prop_dictionary[0][theme_id]["evaluation"] = Evaluation.NO_PREDICTION_5
                 continue
-            dist_predicted_max = max(list(dict_predicted_keys[theme_id].keys()))
-            logging.debug(
-                "max predicted dist proposed for theme_id "
-                + str(theme_id)
-                + " : "
-                + str(dist_predicted_max)
-            )
-            predicted_result = dict_predicted[dist_predicted_max][theme_id]
-            dict_evaluated_result[dist_predicted_max][theme_id] = predicted_result
-            prop_dictionary[dist_predicted_max][theme_id]["formula"] = json.dumps(
-                actual_aligner.get_formula(predicted_result["result"])
-            )
-            prop_dictionary[dist_predicted_max][theme_id][
-                "evaluation"
-            ] = Evaluation.TO_CHECK_4
+            # Add all predicted features so they can be manually checked
+            for dist in dict_predicted_keys[theme_id].keys():
+                predicted_resultset = dict_predicted[dist][theme_id]
+                dict_evaluated_result[dist][theme_id] = predicted_resultset
+                prop_dictionary[dist][theme_id]["formula"] = json.dumps(
+                    actual_aligner.get_formula(predicted_resultset["result"])
+                )
+                prop_dictionary[dist][theme_id]["evaluation"] = Evaluation.TO_CHECK_4
 
     for theme_id, geom in dict_unchanged.items():
         result = {"result": geom}
@@ -365,6 +358,7 @@ def check_equality(
     # TODO: research naar aanduid_id 116448 (equality na 0.5m), 120194 (1m)
     # TODO: research and implementation of following ideas
     # TODO: refine equality comparison, make it more generic
+    # TODO: Add control of OD to equality-comparison (see case aanduid_id 120288)
     # ideas:
     # * If result_diff smaller than 0.x --> automatic update
     # * big polygons: If 'outer ring' has same formula (do net check inner side) -->
@@ -373,9 +367,25 @@ def check_equality(
 
     if base_formula is None or actual_formula is None:
         return False, Evaluation.NO_PREDICTION_5
+    od_alike = False
+    if base_formula["reference_od"] is None and actual_formula["reference_od"] is None:
+        od_alike = True
+    elif base_formula["reference_od"] is None or actual_formula["reference_od"] is None:
+        od_alike = False
+    elif (
+        abs(
+            base_formula["reference_od"]["area"]
+            - actual_formula["reference_od"]["area"]
+        )
+        * 100
+        / base_formula["reference_od"]["area"]
+    ) < threshold_percentage:
+        od_alike = True
+
     if (
         base_formula["reference_features"].keys()
         == actual_formula["reference_features"].keys()
+        and od_alike
     ):
         if base_formula["full"] and base_formula["full"]:
             return True, Evaluation.EQUALITY_FORMULA_GEOM_1
@@ -409,7 +419,7 @@ def check_equality(
                 equal_reference_features = False
         if equal_reference_features:
             return True, Evaluation.EQUALITY_FORMULA_2
-    if base_formula["full"] and base_formula["full"]:
+    if base_formula["full"] and base_formula["full"] and od_alike:
         return True, Evaluation.EQUALITY_GEOM_3
     return False, Evaluation.NO_PREDICTION_5