+++ examples/__init__.py
@@ -1,8 +1,8 @@
 import logging
+from math import ceil
 
 import geopandas as gpd
 import matplotlib.pyplot as plt
-from math import ceil
 
 from brdr.typings import ProcessResult
 from brdr.utils import processresult_to_dicts